Rename PosCsvUtils file
