+++ lib/bloc/pos_catalog/bloc.dart
@@ -4,7 +4,7 @@ import 'dart:convert';
 import 'package:breez/bloc/account/account_model.dart';
 import 'package:breez/bloc/async_actions_handler.dart';
 import 'package:breez/bloc/pos_catalog/actions.dart';
-import 'package:breez/bloc/pos_catalog/pos_csv_exporter.dart';
+import 'package:breez/bloc/pos_catalog/pos_csv_utils.dart';
 import 'package:breez/bloc/pos_catalog/repository.dart';
 import 'package:breez/bloc/pos_catalog/sqlite/repository.dart';
 import 'package:breez/bloc/user_profile/currency.dart';
